2002-08-08  Bradford Hovinen  <hovinen@cis.udel.edu>
	* linbox/vector/subvector.h (Subvector::Subvector): Use Iterator
	() rather than 0

	* linbox/field/param-fuzzy.h (class ParamFuzzy): Use std:: for
	istream and ostream

	* tests/test-scalar-matrix.C (testScalarApply): Make j a size_t

	* tests/test-submatrix.C (testRandomApply): Make i, j, k, l
	size_t's

	* tests/test-subvector.C (testSubvector): Remove unused variable
	i, make z a size_t

	* linbox/vector/const-iterator-type.h: Qualify vector with std::

	* tests/test-subiterator.C (test): Use
	Subiterator<std::vector<int>::iterator>

	* linbox/vector/subiterator.h: Reindent

	* linbox/vector/const-iterator-type.h: Rename from
	constiteratortype.h; reindent

	* tests/test-randiter-nonzero.C: Remove #include <strstream>

+++ trunk/linbox/tests/test-submatrix.C
@@ -69,7 +69,7 @@ static bool testRandomApply (Field                                           &F,
 
 	Vector v, w1(n), w2(n);
 
-	int i, j, k, l;
+	size_t i, j, k, l;
 
 	Blackbox *Ai[9];
 	Blackbox A (F, n * 3, n * 3);
